# backend/docs/swagger.yaml
definitions:
...
  dto.CreateMangaRequest:
    properties:
      artist:
        type: string
      author:
        type: string
      category:
        type: string
      description:
        type: string
      status:
        $ref: '#/definitions/model.MangaStatus'
      tags:
        items:
          type: string
        type: array
      title:
        minLength: 1
        type: string
      type:
        $ref: '#/definitions/model.MangaType'
    required:
    - title
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.MangaResponse:
    properties:
      artist:
        type: string
      author:
        type: string
      category:
        type: string
      cover_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      owner_id:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/model.MangaStatus'
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        $ref: '#/definitions/model.MangaType'
      updated_at:
        type: string
    type: object
...
  dto.PagedMangaResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.MangaResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  dto.UpdateMangaRequest:
    properties:
      artist:
        type: string
      author:
        type: string
      category:
        type: string
      description:
        type: string
      status:
        $ref: '#/definitions/model.MangaStatus'
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        $ref: '#/definitions/model.MangaType'
    type: object
...
  model.MangaStatus:
    enum:
    - ongoing
    - completed
    - hiatus
    type: string
    x-enum-varnames:
    - StatusOngoing
    - StatusCompleted
    - StatusHiatus
  model.MangaType:
    enum:
    - series
    - oneshot
    type: string
    x-enum-varnames:
    - TypeSeries
    - TypeOneshot
info:
  contact: {}
paths:
...
  /mangas:
    get:
      parameters:
      - description: Search title/description
        in: query
        name: q
        type: string
      - description: Filter by status
        enum:
        - ongoing
        - completed
        - hiatus
        in: query
        name: status
        type: string
      - collectionFormat: csv
        description: Filter by tags (AND)
        in: query
        items:
          type: string
        name: tags[]
        type: array
      - description: Filter by author (partial match)
        in: query
        name: author
        type: string
      - description: Filter by artist (partial match)
        in: query
        name: artist
        type: string
      - description: Filter by category (partial match)
        in: query
        name: category
        type: string
      - description: Sort order
        enum:
        - newest
        - oldest
        - title
        in: query
        name: sort
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 24
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PagedMangaResponse'
      summary: List manga
      tags:
      - manga
    post:
      consumes:
      - application/json
      parameters:
      - description: Manga data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMangaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MangaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create manga
      tags:
      - manga
  /mangas/{mangaID}:
    delete:
      parameters:
      - description: Manga ID
        in: path
        name: mangaID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete manga
      tags:
      - manga
    get:
      parameters:
      - description: Manga ID
        in: path
        name: mangaID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MangaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get manga by ID
      tags:
      - manga
    patch:
      consumes:
      - application/json
      parameters:
      - description: Manga ID
        in: path
        name: mangaID
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMangaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MangaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update manga
      tags:
      - manga
...
  /mangas/{mangaID}/cover:
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Manga ID
        in: path
        name: mangaID
        required: true
        type: string
      - description: Cover image (jpeg/png/webp)
        in: formData
        name: cover
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MangaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload manga cover
      tags:
      - manga
...
  /users/{userID}/mangas:
    get:
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PagedMangaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List manga by user
      tags:
      - manga
...
swagger: "2.0"